generate links to scrape
Generates links to scrape by first taking the root url "iss.tamu.edu"
and then gets the html code for that website and filters the href
links in that which are relative - this gives us the links we can crawl
under the same domain while keeping the number of urls short to save on
cost later.

Also contains openssl setup for using with requests. This was not
trivial as latest openssl version is 3.xx which has options that iss
tamu website is not compatible with. Also, the iss tamu website does
not return the entire certificate chain, which causes certificate
validation to fail (as openssl does not automatically download
intermediate certificates). It also does not come with root certificate.
To solve this, a crts directory has been created which has the
certificates which complete the chain for the iss website. openssl
rehash had to be used to correct the naming of the certificates.
This code can and should be used later when getting data from the
websites.